# [parsers] adding check for envelope_keys and validating if they exist